requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
pillow==10.4.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
